upd pathfinding
